MidiFile->BeatsPerMinute into double. Accordingly also the MidiEvent.cs->DeltaTimes and MidiTrack.cs->TotalTime had to be changed.
Changed in SampleSequencer the parameter from uint to double.
private double DeltaTimetoSamples(double DeltaTime)

MidiSequencer synchronized with audio in Unity works pretty nice without any lags.

Removed all possible warnings from the project.

+++ src/CSharpSynth/Midi/MidiTrack.cs
@@ -4,7 +4,7 @@
     {
         //--Variables
         public uint NotesPlayed;
-        public ulong TotalTime;
+        public double TotalTime;
         public byte[] Programs;
         public byte[] DrumPrograms;
         public MidiEvent[] MidiEvents;
